Check for common errors in respirometry data

inspect_data() scans a data frame for specific errors that may affect the use of functions in respR.

inspect_data(
  df,
  time = NULL,
  oxygen = NULL,
  inflow.o2 = NULL,
  outflow.o2 = NULL,
  highlight = TRUE,
  plot = TRUE
)

Arguments

df
data frame. Accepts data frame object of any size.

time
numeric. Defaults to NULL. This is the time data.

oxygen
numeric. Defaults to NULL. This is the dissolved oxygen data.

inflow.o2
numeric. Defaults to NULL. This is inflow oxygen data. Used only for flowthrough respirometry data.

outflow.o2
numeric. Defaults to NULL. This is outflow oxygen data. Used only for flowthrough respirometry data.

highlight
logical. Defaults to TRUE. Prints location (row #) of errors detected by the function.

plot
logical. Defaults to TRUE. Produces plots for quick visual diagnostics.

Value

A list object of class adjust_rate.

Details

PLEASE NOTE: the inspect_data function is deprecated. It will not be updated, and will be removed in a future update to respR. Please use the new inspect function instead.

Data checks include:

A test for NA/NaN inputs. * A test for numeric data. * A test for sequential time data. * A test for duplicate time data. * A test for evenly-spaced time data.

Once data checks are complete, the function produces a list object which may be directly loaded into calc_rate(), calc_rate.bg(), calc_rate.ft(), and auto_rate() for further analyses.

Examples

data("sardine.rd")
inspect_data(sardine.rd)

#> Warning: inspect_data has been deprecated. 
#> It will not be updated and will be removed in a future version of respR. 
#> Please use the `inspect` function instead.

#> Error in `$<-.data.frame`(x, name, value): replacement has 0 rows, data has 7513

data("urchins.rd")
inspect_data(urchins.rd, 1, 5, highlight = FALSE)

#> Warning: inspect_data has been deprecated. 
#> It will not be updated and will be removed in a future version of respR. 
#> Please use the `inspect` function instead.

#> a :  num [1:271] 7.86 7.87 7.89 7.9 7.87 7.82 7.84 7.84 7.86 7.81 ...
#> b :  num [1:271] 7.86 7.79 7.7 7.68 7.64 7.69 7.69 7.66 7.65 7.64 ...
#> b1 :  num [1:271] 7.9 7.89 7.9 7.89 7.9 7.9 7.92 7.91 7.86 7.89 ...
#> b2 :  num [1:271] 7.7 7.7 7.69 7.69 7.68 7.67 7.68 7.69 7.69 7.66 ...
#> c :  num [1:271] 7.64 7.6 7.6 7.6 7.6 7.61 7.56 7.59 7.59 7.59 ...
#> d :  num [1:271] 7.65 7.71 7.7 7.72 7.67 7.61 7.62 7.63 7.63 7.57 ...
#> e :  num [1:271] 7.87 7.87 7.9 7.92 7.9 7.88 7.85 7.83 7.82 7.85 ...
#> f :  num [1:271] 7.74 7.72 7.72 7.74 7.73 7.7 7.7 7.69 7.64 7.64 ...
#> g :  num [1:271] 7.62 7.61 7.61 7.62 7.59 7.6 7.6 7.6 7.62 7.61 ...
#> h :  num [1:271] 7.65 7.66 7.63 7.66 7.65 7.65 7.67 7.69 7.65 7.66 ...
#> i :  num [1:271] 7.96 7.97 7.98 7.97 7.95 7.94 7.95 7.96 7.94 7.96 ...
#> j :  num [1:271] 7.75 7.72 7.72 7.72 7.71 7.7 7.69 7.7 7.67 7.66 ...
#> k :  num [1:271] 7.72 7.71 7.69 7.7 7.66 7.63 7.65 7.65 7.69 7.67 ...
#> l :  num [1:271] 7.71 7.71 7.77 7.77 7.76 7.72 7.73 7.74 7.74 7.76 ...
#> m :  num [1:271] 7.87 7.89 7.89 7.89 7.86 7.86 7.84 7.83 7.85 7.85 ...
#> n :  num [1:271] 7.61 7.61 7.65 7.67 7.62 7.61 7.62 7.64 7.64 7.62 ...
#> o :  num [1:271] 6.96 6.96 6.97 6.96 6.95 6.94 6.93 6.93 6.9 6.93 ...
#> p :  num [1:271] 7.04 7.01 7.05 7.09 7 6.99 7.02 7.03 7 6.97 ...
#> time.min :  num [1:271] 0 0.2 0.3 0.5 0.7 0.8 1 1.2 1.3 1.5 ...
#> ---
#>                     Score  
#> No NA/NaN in time   PASS   
#> No NA/NaN in Oxygen PASS   
#> Sequential Time     PASS   
#> Non-duplicated Time PASS   
#> Evenly-spaced Time  FAIL***

#> Warning: Time data is irregular. Subsetting by `row` may result in irregular durations.

#> New dataframe generated.

#> Error in roll_lm(matrix(df[[1]]), matrix(df[[2]]), win): could not find function "roll_lm"